+++ tests/fixtures/projects/async-await/nwb.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  type: 'react-app',
+  babel: {
+    stage: 0,
+    optional: ['runtime']
+  }
+}